requirements.txt
sqlalchemy==2.0.43
aiogram==3.22.0
python-decouple==3.8